moved components dir to /src
