+++ src/app/layout.js
@@ -1,6 +1,6 @@
 import { Geist, Geist_Mono } from "next/font/google";
 import "./globals.css";
-import ClientLayout from "./components/ClientLayout"; // 👈 import the wrapper
+import ClientLayout from "@/components/ClientLayout"; // 👈 import the wrapper
 
 const geistSans = Geist({
   variable: "--font-geist-sans",
